Revert "sched: cass: Compare CPU overload on wake-up latency check."
As with my other additions to the pipeline, experimental.
The overloading comparison over headroom comparison seems to yield poor results.
While the UI is overall smooth, scrolling the launcher's app drawer has noticeable lag,
which stops when the CPU is busy, due to all cores being awake and active.

For now, lets go back to the old method.
As overloading does not seem to give (many) good results.

This reverts commit 901b6720113bb5fb98023cc120fb8cf3bf58505b.